requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
async-timeout
attrs
backoff
bcrypt
build
cachetools
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
coloredlogs
cryptography
dataclasses-json
Deprecated
distro
durationpy
ecdsa
exceptiongroup
fastapi
filelock
flatbuffers
frozenlist
fsspec
google-auth
googleapis-common-protos
greenlet
grpcio
h11
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
iniconfig
jiter
jsonpatch
jsonpointer
kubernetes
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
markdown-it-py
marshmallow
mdurl
mmh3
monotonic
mpmath
multidict
mypy-extensions
numpy
oauthlib
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
oso
overrides
packaging
passlib
pluggy
posthog
propcache
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
PyJWT
pypdf
PyPika
pyproject_hooks
pytest
python-dateutil
python-dotenv
python-jose
python-multipart
PyYAML
regex
requests
requests-oauthlib
requests-toolbelt
rich
rsa
shellingham
six
sniffio
SQLAlchemy
starlette
sympy
tenacity
tiktoken
tokenizers
tomli
tqdm
typer
typing-inspect
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
websocket-client
websockets
wrapt
yarl
zipp
zstandard
